docker: add apache ports, init functions

+++ resources/install_composer.sh
@@ -1,9 +1,9 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 echo "Begin installation of composer"
 wget https://getcomposer.org/installer -O composer-setup.php
 php composer-setup.php
 php -r "unlink('composer-setup.php');"
-sudo rm /usr/local/bin/composer
+[[ -f /usr/local/bin/composer ]] && sudo rm /usr/local/bin/composer || true
 sudo mv composer.phar /usr/local/bin/composer
 echo "End installation of composer"
