+++ resources/install_nodejs.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 installVer='20' 	#NodeJS major version to be installed
 minVer='20'	      #min NodeJS major version to be accepted
@@ -145,9 +145,9 @@ else
     echo "Utilisation du dépot officiel"
     NODE_MAJOR=$installVer
     sudo mkdir -p /etc/apt/keyrings
-    sudo rm /etc/apt/keyrings/nodesource.gpg
+    [[ -f /etc/apt/keyrings/nodesource.gpg ]] && sudo rm /etc/apt/keyrings/nodesource.gpg || true
     curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
-    sudo rm /etc/apt/sources.list.d/nodesource.list
+    [[ -f /etc/apt/sources.list.d/nodesource.list ]] && sudo rm /etc/apt/sources.list.d/nodesource.list || true
     echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | sudo tee /etc/apt/sources.list.d/nodesource.list
     sudo apt-get update
     sudo DEBIAN_FRONTEND=noninteractive apt-get install -y nodejs
@@ -232,7 +232,7 @@ fi
 fi
 
 # on nettoie la priorité nodesource
-sudo rm -f /etc/apt/preferences.d/nodesource &>/dev/null
+[[ -f /etc/apt/preferences.d/nodesource ]] && sudo rm -f /etc/apt/preferences.d/nodesource &>/dev/null || true
 
 # on remet deb-multimedia si on l'a désactivé avant
 if [ -f /etc/apt/sources.list.d/deb-multimedia.list.disabled ]; then
